Entity 작성 - User 테이블은 user 가 h2 예약어로 인해 실제 테이블명은 user_table 로 대체

+++ src/main/java/com/chaw/hhplus_tdd_lecture/domain/lecture/entity/ApplicationDetail.java
@@ -14,11 +14,11 @@ public class ApplicationDetail {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
-    private Integer userId; // FK
+    private Long userId; // FK
 
-    private Integer lectureItemId; // FK
+    private Long lectureItemId; // FK
 
     private LocalDateTime registrationDate;
 }
